More chunkInput into common
